Fixed a few conversion warnings and added a spin to the ship, to see if the plane was at an angle that wouldn't be viewable, not the case.

+++ Direct3D 11/Camera.cpp
@@ -18,8 +18,8 @@ Camera::Camera( const Window & Win )
 	orientation( XMFLOAT4( 0.0f, 0.0f, 0.0f, 1.0f ) ),
 	vertical_fov( XMConvertToRadians( 70 ) )
 {
-	viewport.Width = Win.GetWidth( );
-	viewport.Height = Win.GetHeight( );
+	viewport.Width = static_cast<float>( Win.GetWidth( ) );
+	viewport.Height = static_cast<float>( Win.GetHeight( ) );
 	viewport.TopLeftX = 0.0f;
 	viewport.TopLeftY = 0.0f;
 	viewport.MinDepth = 0.1f;
